Milter: Abort upon EHLO/HELO midway conversation.
If you have an SMTP conversation where the client e.g. does:
HELO localhost
Mail from: johndoe
HELO localhost

Then libmilter seems to lose its context and returns an empty one
instead. Because we need some context (like client ip), we abort
the conversation with a 4XX error. Hoping that if it's a decent
MTA it's a one-time mistake, and will retry later.

Postfix behaves the same way with a client that sends it nothing
but junk.

We do try to display the ip address in the error, but with the
versions of libmilter I tested this with it also returned an
empty string.