Store body in chunks, rather than as longtext.
This was causing problems with max_allowed_packet, this should prevent this. In theory, chunks will never be
larger than 65535 bytes. But, according to the milter documentation these can contain nullbytes. Given that
a nullbyte will be escaped by another nullbyte by mysql, the Blob data type doesn't suffice, and we need
to use Mediumblob.