Schema change: change address field from varbinary(16) to varchar(45)
Turned out we were selecting ip's from the session table to that of the
greylist table. However, in the session table they are - currently -
administered in plain text. The idea was to store them in binary notation
in the greylist table. However, due to a lack of a ntop() function in
mysql that will, for now, require too much effort. As such, the ip type
in the greylist table was changed to reflect that of the session table