Set mysql isolation level to READ-UNCOMMITTED
Almost all our data is not based off a single row, but is the result of new rows for each message.
As such, we don't really care about ACIDness, but we do care about preventing deadlocks. As such,
we'll default to setting the isolation level to a rather low level.

This may be reverted or made configurable later on